Restructured files and added update functionality to cocktails.
