+++ lib/all-recipe-books.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const client = require('../lib/client');
+const client = require('./client');
 
 function viewAllRecipeBooks(request, response) {
     let sql = "SELECT * FROM cocktail_lists WHERE user_id=$1;"
