+++ lib/cocktailHandler.js
@@ -2,7 +2,7 @@
 
 const superagent = require('superagent')
 const cocktailConstructor = require('./cocktailConstructor.js');
-const client = require('../lib/client');
+const client = require('./client');
 
 module.exports = {
   getCocktailsByBase: function(request, response) {
@@ -162,22 +162,7 @@ module.exports = {
     let safeValues1 = [request.body.title, request.body.instructions, request.body.ingredients, id]
     client.query(sql1, safeValues1)
         .then(results => {
-          // let sql2 = `DELETE FROM cocktails WHERE id = $1;`;
-    
-          // let safeValues2 = [id];
-        
-          // client.query(sql2, safeValues2)
-          //   .then(() => {
-          //     let sql1 = "SELECT * FROM cocktail_lists WHERE cocktail_list_id=$1;"
-          //     let safeValues1 = [results.rows[0].cocktail_list]
-          //     client.query(sql1, safeValues1)
-          //       .then(results => {
-          //         response.redirect('/recipe-book/' + results.rows[0].list_name);
-          //     })
-          //   })
-          //   .catch(error => {
-          //     handleError(error, response);
-          //   });
+          response.redirect('/recipe-book/' + request.body.currentList)
       })
     
   }
