+++ lib/loginHandler.js
@@ -1,7 +1,7 @@
 'use strict';
 
 
-const client = require('../lib/client');
+const client = require('./client');
 
 function handleLogin(request, response) {
     console.log(request.body.userName)
