+++ lib/recipe-book.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const client = require('../lib/client');
+const client = require('./client');
 
 
 function renderRecipeBookPage(request, response) {
@@ -9,19 +9,19 @@ function renderRecipeBookPage(request, response) {
     let safeValues = [global.username, request.params.id]
     client.query(sql, safeValues)
       .then(results => {
-        console.log("LIST SELECT RESULTS", results.rows)
+        // console.log("LIST SELECT RESULTS", results.rows)
         let sql2 = "SELECT * FROM cocktails WHERE cocktail_list=$1;";
         let safeValues2 = [results.rows[0].cocktail_list_id]
         
-        let listName = results.rows[0].list_name
-        console.log("LIST NAME", listName)
+        let list = results.rows[0]
+        console.log("LIST", list)
 
         client.query(sql2, safeValues2)
           .then(results => {
             let cocktails = results.rows;
-            console.log(cocktails[0])
-            console.log("username:", global.username)
-            response.render('database/recipe-book', { cocktailArray: cocktails, username: global.username, listName: listName})
+            // console.log(cocktails[0])
+            // console.log("username:", global.username)
+            response.render('database/recipe-book', { cocktailArray: cocktails, username: global.username, list: list})
           })
       })
